MNT: Write repodata into a temp dir then move
Currently conda index does an in-place rewrite of repodata.json.  If
this is done while someone is trying to conda install from that channel
then there is a reasonable chance that they can encounter a bad
repodata.json file given enough traffic. This PR write new repodata to a
temporary location and then will use the filesystem to atomically move
it to the platform directory.